Feat : Endpoints and DTO implemented

+++ src/main/java/fr/hatclic/chatop/service/MessagesService.java
@@ -1,10 +1,8 @@
 package fr.hatclic.chatop.service;
 
 import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import fr.hatclic.chatop.model.Messages;
 import fr.hatclic.chatop.repository.MessagesRepository;
 
@@ -15,10 +13,15 @@ public class MessagesService {
 	private MessagesRepository messagesRepository;
 
 	public Messages createMessage(Messages message) {
+		if (messageIsNull(message))
+			throw new Error();
 		return messagesRepository.save(message);
 	}
 
 	public Optional<Messages> findMessageById(final Long id) {
+		if (id == null) {
+			throw new Error();
+		}
 		return messagesRepository.findById(id);
 	}
 
@@ -27,11 +30,24 @@ public class MessagesService {
 	}
 
 	public Messages updateMessage(Messages message) {
+		if (messageIsNull(message) || message.getId() == null)
+			throw new Error();
 		return messagesRepository.save(message);
 	}
 
 	public void deleteMessageById(final Long id) {
+		if (id == null) {
+			throw new Error();
+		}
 		messagesRepository.deleteById(id);
+	}
+
+	public boolean messageIsNull(Messages message) {
+		if (message == null || message.getRental_id() == null || message.getUser_id() == null
+				|| message.getMessage().trim().length() == 0) {
+			return true;
+		}
+		return false;
 	}
 
 }
