+++ src/main/java/fr/hatclic/chatop/service/RentalsService.java
@@ -4,7 +4,6 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import fr.hatclic.chatop.model.Rentals;
 import fr.hatclic.chatop.repository.RentalsRepository;
 
@@ -15,10 +14,15 @@ public class RentalsService {
 	private RentalsRepository rentalsRepository;
 
 	public Rentals createRental(Rentals rental) {
+		if (rentalIsNull(rental))
+			throw new Error();
 		return rentalsRepository.save(rental);
 	}
 
 	public Optional<Rentals> findRentalById(final Long id) {
+		if (id == null) {
+			throw new Error();
+		}
 		return rentalsRepository.findById(id);
 	}
 
@@ -27,11 +31,25 @@ public class RentalsService {
 	}
 
 	public Rentals updateRental(Rentals rental) {
+		if (rentalIsNull(rental) || rental.getId() == null)
+			throw new Error();
 		return rentalsRepository.save(rental);
 	}
 
 	public void deleteRentalById(final Long id) {
+		if (id == null) {
+			throw new Error();
+		}
 		rentalsRepository.deleteById(id);
+	}
+
+	public boolean rentalIsNull(Rentals rental) {
+		if (rental == null || rental.getName().trim().length() == 0 || rental.getSurface() == 0.0
+				|| rental.getPrice() == 0.0 || rental.getDescription().trim().length() == 0
+				|| rental.getOwner_id() == null) {
+			return true;
+		}
+		return false;
 	}
 
 }
